+++ docs/changes.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+Version 0.5.1
+'''''''''''''
+
+To be released.
+
+
 Version 0.5.0
 '''''''''''''